refactor(shadow-reading): split panel widgets into presentation/widgets/ (#180)
Extract four self-contained widgets/painter from the 1372-line
shadow_reading_panel.dart into presentation/widgets/, matching the layout
used by 8 other features (auth, hotkeys, library, player, ...):

- shadow_record_fab.dart          (ShadowRecordFab + ShadowRecordRingPainter)
- shadow_reading_toolbar_row.dart (ShadowReadingToolbarRow)
- shadow_recording_caption.dart   (ShadowRecordingCaptionRow)
- shadow_takes_toolbar_actions.dart (ShadowTakesToolbarActions +
  confirmShadowDeleteTake + popup-menu value tokens)

This is a mechanical, behavior-preserving move: only the leading underscore
was dropped on the renamed classes; no logic changed. Pure helper
_formatSecsOneDecimal moved into shadow_recording_caption.dart (its sole
consumer). The public ShadowReadingPanel API is unchanged.

Panel: 1372 -> 779 lines. All gates green: flutter analyze, flutter test
(731 passing), dart format --set-exit-if-changed.

The proposed application/shadow_reading_recording_controller.dart extraction
is intentionally deferred — see PR description for rationale.
